flask_mongo_assignment/backend/requirements.txt
flask==3.0.3
pymongo==4.8.0
flask-cors==4.0.0
python-dotenv==1.0.1
